Rename package and repo (#23) (#25)
* Rename package and repo (#23)

* Update composer.json and InstallComposerPackagesCommand.php debs. (#17)

* Update composer.json and InstallComposerPackagesCommand.php files.

The changes were made to update the required versions of certain packages and add new packages to the composer.json file. In InstallComposerPackagesCommand.php, a new method was added to execute shell commands and display information and error messages.

* Fix styling

---------

Authored-by: Joshua Mitchener <thejmitchener@gmail.com>

* Fix styling

* Fix styling

* Update test matrix to use single version for Testbench

Simplify the version specification for Testbench in the CI workflow. This ensures better clarity and consistency in dependency definitions.

* Fix styling

* Fix styling

* Fix styling

* Fix styling

* Update dependencies and refine GitHub Actions workflow

Allow compatibility with orchestra/testbench 9.0 in composer.json. Simplify the CI matrix configuration by adjusting Laravel and Testbench version exclusions and removing redundant OS specification.

* Fix styling

* Fix styling

* Update CI workflow to support multiple OS environments

Refactored the GitHub Actions workflow to use matrix strategies for OS, enabling tests to run on Ubuntu. Added a timeout for jobs and improved matrix naming for better clarity. Adjusted matrix inclusions and configurations to streamline test coverage.

* Fix styling

* Fix styling

* Update matrix to test only Laravel 11 and PHP 8.3

Simplify the testing matrix by focusing on Laravel 11 and PHP 8.3. This reduces complexity and aligns tests with the latest supported versions, ensuring compatibility with the targeted stack. Updated testbench to version 9 for consistency.

* Fix styling

* Fix styling

* Expand test matrix to include PHP 8.2 and Laravel 10.

This update adjusts the GitHub Actions workflow to test against PHP 8.2 and Laravel 10, alongside the existing versions. It ensures compatibility with a broader range of environments and improves overall test coverage.

* Fix styling

* Fix styling

---------

Co-authored-by: Joshua Mitchener <thejmitchener@gmail.com>